requirements.txt
# Discord Bot Core
discord.py>=2.3.0
openai>=1.40.0
python-dotenv>=1.0.0

# Database
motor>=3.3.0
pymongo[srv]>=4.6.0
dnspython>=2.5.0

# Web & HTTP
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# AI & ML
runware>=0.4.33
tiktoken>=0.7.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Document Processing
pypdf>=4.0.0
Pillow>=10.0.0

# Scheduling & Time
APScheduler>=3.10.0
tzlocal>=5.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
ruff>=0.3.0

# Monitoring & Logging (Optional)
# sentry-sdk>=1.40.0  # Uncomment for error monitoring
# python-json-logger>=2.0.0  # Uncomment for structured logging